refactor(LottoService):getLotto, getLottoRank 메서드 분리 dquote> dquote> 이전 : LottoService에서 모두 관리 dquote> 변경 : LottoUtil 클래스로 분리 dquote> dquote>
