+++ src/test/kotlin/lottoTest/LottoServiceTest.kt
@@ -18,76 +18,6 @@ class LottoServiceTest {
     }
 
     @Test
-    @DisplayName("지난 주 당첨 번호와 보너스 볼을 입력받으면 복권의 당첨 등수(enum 멤버)를 반환한다")
-    fun t3() {
-        val lotto = Lotto(listOf(1, 2, 3, 4, 5, 6))
-        val winningLotto = Lotto(listOf(1, 2, 3, 4, 5, 6))
-        val bonus = 7
-        val result = lottoService.getLottoRank(lotto, winningLotto, bonus)
-        assertThat(result).isEqualTo(Rank.FIRST)
-    }
-
-    @Test
-    @DisplayName("지난 주 당첨 번호와 보너스 볼을 입력받으면 복권의 당첨 등수(enum 멤버)를 반환한다")
-    fun t3_1() {
-        val lotto = Lotto(listOf(1, 2, 3, 4, 5, 6))
-        val winningLotto = Lotto(listOf(1, 2, 3, 4, 5, 8))
-        val bonus = 7
-        val result = lottoService.getLottoRank(lotto, winningLotto, bonus)
-        assertThat(result).isEqualTo(Rank.THIRD)
-    }
-
-    @Test
-    @DisplayName("지난 주 당첨 번호와 보너스 볼을 입력받으면 복권의 당첨 등수(enum 멤버)를 반환한다")
-    fun t3_2() {
-        val lotto = Lotto(listOf(1, 2, 3, 4, 5, 6))
-        val winningLotto = Lotto(listOf(1, 2, 3, 4, 5, 7))
-        val bonus = 7
-        val result = lottoService.getLottoRank(lotto, winningLotto, bonus)
-        assertThat(result).isEqualTo(Rank.SECOND)
-    }
-
-    @Test
-    @DisplayName("지난 주 당첨 번호와 보너스 볼을 입력받으면 복권의 당첨 등수(enum 멤버)를 반환한다")
-    fun t3_2_1() {
-        val lotto = Lotto(listOf(1, 2, 3, 4, 5, 6))
-        val winningLotto = Lotto(listOf(1, 2, 3, 4, 5, 7))
-        val bonus = 5
-        val result = lottoService.getLottoRank(lotto, winningLotto, bonus)
-        assertThat(result).isEqualTo(Rank.THIRD)
-    }
-
-    @Test
-    @DisplayName("지난 주 당첨 번호와 보너스 볼을 입력받으면 복권의 당첨 등수(enum 멤버)를 반환한다")
-    fun t3_3() {
-        val lotto = Lotto(listOf(1, 2, 3, 4, 5, 6))
-        val winningLotto = Lotto(listOf(1, 2, 3, 4, 8, 9))
-        val bonus = 7
-        val result = lottoService.getLottoRank(lotto, winningLotto, bonus)
-        assertThat(result).isEqualTo(Rank.FOURTH)
-    }
-
-    @Test
-    @DisplayName("지난 주 당첨 번호와 보너스 볼을 입력받으면 복권의 당첨 등수(enum 멤버)를 반환한다")
-    fun t3_4() {
-        val lotto = Lotto(listOf(1, 2, 3, 4, 5, 6))
-        val winningLotto = Lotto(listOf(1, 2, 3, 10, 8, 9))
-        val bonus = 7
-        val result = lottoService.getLottoRank(lotto, winningLotto, bonus)
-        assertThat(result).isEqualTo(Rank.FIFTH)
-    }
-
-    @Test
-    @DisplayName("지난 주 당첨 번호와 보너스 볼을 입력받으면 복권의 당첨 등수(enum 멤버)를 반환한다")
-    fun t3_5() {
-        val lotto = Lotto(listOf(1, 2, 3, 4, 5, 6))
-        val winningLotto = Lotto(listOf(1, 2, 11, 10, 8, 9))
-        val bonus = 7
-        val result = lottoService.getLottoRank(lotto, winningLotto, bonus)
-        assertThat(result).isEqualTo(Rank.MISS)
-    }
-
-    @Test
     @DisplayName("복권을 구매한 숫자만큼 복권의 당첨 등수를 반환한다")
     fun t4() {
         val manyLotto =
